+++ app/src/main/java/com/baghdad/novix/di/LocalDataSourceModule.kt
@@ -7,13 +7,13 @@ import com.baghdad.local_datasource.LocalMovieDataSourceImpl
 import com.baghdad.local_datasource.LocalRecentlyViewedDataSourceImpl
 import com.baghdad.local_datasource.LocalSearchDataSourceImpl
 import com.baghdad.local_datasource.LocalTvShowDataSourceImpl
-import com.baghdad.local_datasource.roomDB.dao.ActorDao
-import com.baghdad.local_datasource.roomDB.dao.GenreDao
-import com.baghdad.local_datasource.roomDB.dao.MovieDao
-import com.baghdad.local_datasource.roomDB.dao.RecentSearchDao
-import com.baghdad.local_datasource.roomDB.dao.RecentlyViewedDao
-import com.baghdad.local_datasource.roomDB.dao.TvShowDao
-import com.baghdad.local_datasource.roomDB.database.NovixDatabase
+import com.baghdad.local_datasource.database.dao.ActorDao
+import com.baghdad.local_datasource.database.dao.GenreDao
+import com.baghdad.local_datasource.database.dao.MovieDao
+import com.baghdad.local_datasource.database.dao.RecentSearchDao
+import com.baghdad.local_datasource.database.dao.RecentlyViewedDao
+import com.baghdad.local_datasource.database.dao.TvShowDao
+import com.baghdad.local_datasource.NovixDatabase
 import com.baghdad.repository.datasource.local.LocalActorDataSource
 import com.baghdad.repository.datasource.local.LocalMovieDataSource
 import com.baghdad.repository.datasource.local.LocalRecentSearchDataSource
